Rudimentary VA-API VPP support to softhddevice v3
This is a second version of VA-API VPP patch for softhddevice. Some changes over the previous version:
 * Implemented dynamic temporal reference handling
 * Picture controls (brightness, contrast etc.) now implemented
 * Fixes deinterlacing when playing bottom-field-first material
 * Still more simplifications to postprocessing setup code

This patchs adds support for Intel VA-API VPP filters to softhddevice. It is expected that many things that used to work are
now broken (progressive images, different frame rates, etc.) or that the entire patch does not work well enough for
production use.

The patch currently forces hardware accelerated Motion Compensated deinterlace algorithm and video denoise filters on. In
addition there is support for changing picture controls from the plugin menu.

Unfortunately the current stable libva-intel-driver has buggy implementation of va-api postprocessing surface handling.
There is a staging version of the driver where the most of the bugs have been fixed
(https://github.com/gbeauchesne/libva-intel-driver/tree/17.vpp.vebox).
This patch was developed against the staging driver. If used against stable libva-intel-driver the image stutters or may get
stuck and green surfaces are occasionally rendered as postprocessing result producing flickering video.

Some additional TODO items for the patch:
 * Dynamic adjusting of deinterlace mode (bob, weave, MotionComp, MotionAdaptive) via softhddevice menu
 * Adjusting of denoise strength via softhddevice menu
 * Providing the picture control limits to the plugin menu (e.g. hue range from -180 to 180)
 * Error handling in case certain vaapi features are not available
 * Testing / fixing progressive input
 * Testing / fixing various framerate and display refreshrate combinations
 * Stricter video pipeline capability checking with filter combinations

The patch was/is tested and developed on:
 * ffmpeg 1.2.6
 * libva 1.3.1
 * libva-intel-driver 1.3.3.pre1 (17.vpp.vebox-r1-8-gad3f8b5)
 + Intel D54250WYK NUC kit (Haswell i5)

Signed-off-by: Antti Seppälä <a.seppala@gmail.com>