Initialize surface flags after filter configuration
From: TomJoad

The Filter params are set to the config-value only after changing the
value in the setup for the resolution of the current channel. At next
reboot and when switching channels at least the denoise-Filter is always
initialized with the default value (0.5). This can be fixed by calling
VaapiInitSurface(decoder) in VaapiSetupVideoProcessing after looping
through the enabled filters.

Many thanks to user TomJoad from vdr-portal.de for the patch