Ensure reference surfaces are ready before postprocessing
This commit refactors the surface handling before committing into postprocessing.

The idea here is to query the driver about the number of needed reference surfaces beforehand and skip deinterlacing if
there are not enough of them. Reference surfaces are also checked for correctness and readiness and deinterlacer is skipped
if any of the checks fail.

This new logic avoids feeding invalid surfaces to driver as references so this commit fixes #99
This may also have a positive effect on #79