Switch over to ffmpeg audio demuxer
Audio packet handling is tricky; they are decoded by the same thread that
feeds them. If the thread is blocked waiting for input too long we get vdr
complaining about TS packet not accepted in tranfer mode and playback gets
broken.

In other words we cannot wait for more input so if buffer gets empty we run
into trouble.

The code needs to ensure that this never happens (old code tries to do this).

Better alternative would be to move audio decoder into audio playback thread.
This is the approach that video decoder already has.